Add check_id and README file, code formatting
check_id for product_id in add/delete basket item

+++ lib/model/storage_item.rb
@@ -10,7 +10,7 @@ module Shop
 
     private
     def check_positive_numeric(var_to_check)
-      raise ArgumentError unless var_to_check.is_a?(Numeric) && var_to_check > 0
+      raise ArgumentError unless var_to_check.is_a?(Integer) && var_to_check > 0
       var_to_check
     end
   end
